Bad initial point.
Set up the mvc servlet to map to the url resolver.

Set up the web.xml to map the servlet and have a welcome file.

Added dependencies to the pom for rendering the jsp

Moved the jsp.  Not sure if that does anything.

Added the spring boot class to the project.